refactor: decompose NavigationPrototype into modular components
Extract 1154-line monolith into components/flow/ directory:
- MiniSprite, BloomOverlay, HeroSection, TraitSelector, WaterfallAlbum, IdleField
- Shared types (flowTypes.ts), helpers (flowHelpers.ts), keyframes (flow.css)
- NavigationPrototype.tsx becomes thin orchestrator (~190 lines)

Closes #19
