+++ components/flow/IdleField.tsx
@@ -1,0 +1,57 @@
+import React, { useMemo } from 'react';
+import { ImageNode } from '../../types';
+import { seededRandom } from './flowHelpers';
+import MiniSprite from './MiniSprite';
+
+interface IdleFieldProps {
+    images: ImageNode[];
+    onSelect: (img: ImageNode, rect: DOMRect) => void;
+    canvasW: number;
+    canvasH: number;
+}
+
+const IdleField: React.FC<IdleFieldProps> = ({ images, onSelect, canvasW, canvasH }) => {
+    const nodes = useMemo(() => {
+        const count = Math.min(images.length, 36);
+        const cols = Math.ceil(Math.sqrt(count * (canvasW / canvasH)));
+        const rows = Math.ceil(count / cols);
+        const cellW = canvasW / (cols + 1);
+        const cellH = canvasH / (rows + 1);
+        return images.slice(0, count).map((img: ImageNode, i: number) => ({
+            image: img,
+            x: cellW * ((i % cols) + 1) + (seededRandom(img.id + 'ix') - 0.5) * cellW * 0.4,
+            y: cellH * (Math.floor(i / cols) + 1) + (seededRandom(img.id + 'iy') - 0.5) * cellH * 0.3,
+        }));
+    }, [images, canvasW, canvasH]);
+
+    return (
+        <div className="fixed inset-0">
+            {nodes.map(({ image, x, y }) => {
+                const size = 56 + seededRandom(image.id + 'sz') * 20;
+                const breatheDur = 5 + seededRandom(image.id + 'bd') * 6;
+                const breatheDel = seededRandom(image.id + 'bl') * 4;
+                return (
+                    <div key={image.id}
+                        className="absolute cursor-pointer hover:scale-110 transition-transform duration-300"
+                        style={{
+                            left: x - size / 2, top: y - size / 2,
+                            animation: `drift ${breatheDur}s ease-in-out ${breatheDel}s infinite`,
+                        }}
+                        onClick={(e) => {
+                            const rect = (e.currentTarget as HTMLElement).getBoundingClientRect();
+                            onSelect(image, rect);
+                        }}>
+                        <MiniSprite image={image} size={size} />
+                    </div>
+                );
+            })}
+            <div className="absolute bottom-12 left-1/2 -translate-x-1/2 z-40 text-center">
+                <p className="text-zinc-400 text-sm" style={{ fontFamily: 'Caveat, cursive' }}>
+                    Tap a sprite to begin
+                </p>
+            </div>
+        </div>
+    );
+};
+
+export default IdleField;
