+++ components/flow/index.ts
@@ -1,0 +1,2 @@
+export { default } from './NavigationPrototype';
+export type { FlowPhase, ScoredImage, TrailPoint, AlbumImage } from './flowTypes';
